addressed reviewer concerns, made lexical scoring more selective
