+++ scripts/attacklog_correlator/VERSION
@@ -1,0 +1,1 @@
+0.1.0
