+++ scripts/attacklog_correlator/src/attacklog_correlator/__init__.py
@@ -1,0 +1,3 @@
+"""Attack log correlator package."""
+
+__version__ = "0.1.0"
