+++ scripts/attacklog_correlator/src/attacklog_correlator/__main__.py
@@ -1,0 +1,5 @@
+"""Run the attack log correlator with `python -m attacklog_correlator`."""
+
+from .cli import main
+
+raise SystemExit(main())
